requirements.txt
# Core Backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12

# Database
alembic==1.13.3

# ML and Embeddings (LOCAL - NO GEMINI for embeddings!)
sentence-transformers==3.1.1
scikit-learn==1.5.2
xgboost==2.0.3
lightgbm==4.5.0
imbalanced-learn==0.12.3
numpy==1.26.4
pandas==2.2.3
faiss-cpu==1.9.0
joblib==1.4.2

# Gemini API (ONLY for response generation)
google-generativeai==0.8.3

# Web Apps
streamlit==1.39.0
plotly==5.24.1
altair==5.4.1

# Utilities
requests==2.32.3
tqdm==4.66.5
